chore: added missing data elements for the unattended SmartDocuments API; renamed some code
