+++ src/main/kotlin/net/atos/client/smartdocuments/model/document/SmartDocumentsDocumentRequestDeclarations.kt
@@ -2,7 +2,7 @@
  * SPDX-FileCopyrightText: 2024 Lifely
  * SPDX-License-Identifier: EUPL-1.2+
  */
-package net.atos.client.smartdocuments.model.wizard
+package net.atos.client.smartdocuments.model.document
 
 import jakarta.json.bind.annotation.JsonbProperty
 import net.atos.zac.documentcreation.model.Data
@@ -30,7 +30,9 @@ data class WizardRequest(
     @field:JsonbProperty("SmartDocument")
     val smartDocument: SmartDocument,
 
+    @field:JsonbProperty("registratie")
     val registratie: Registratie,
 
+    @field:JsonbProperty("data")
     val data: Data
 )
