+++ src/main/kotlin/net/atos/client/smartdocuments/model/document/SmartDocumentsDocumentResponseDeclarations.kt
@@ -2,7 +2,7 @@
  * SPDX-FileCopyrightText: 2024 Lifely
  * SPDX-License-Identifier: EUPL-1.2+
  */
-package net.atos.client.smartdocuments.model.wizard
+package net.atos.client.smartdocuments.model.document
 
 import jakarta.json.bind.annotation.JsonbProperty
 import nl.lifely.zac.util.NoArgConstructor
@@ -15,13 +15,24 @@ data class UnattendedResponse(
 
 @NoArgConstructor
 data class File(
+    @field:JsonbProperty("filename")
     var fileName: String? = null,
-    var document: Document? = null
+
+    @field:JsonbProperty("document")
+    var document: Document? = null,
+
+    @field:JsonbProperty("outputFormat")
+    var outputFormat: String? = null
 )
 
-class Document
+@NoArgConstructor
+class Document(
+    @field:JsonbProperty("data")
+    var data: String? = null
+)
 
 @NoArgConstructor
 data class WizardResponse(
+    @field:JsonbProperty("ticket")
     var ticket: String? = null
 )
