+++ src/main/kotlin/net/atos/client/smartdocuments/model/template/SmartDocumentsTemplateDeclarations.kt
@@ -2,7 +2,7 @@
  * SPDX-FileCopyrightText: 2024 Lifely
  * SPDX-License-Identifier: EUPL-1.2+
  */
-package net.atos.client.smartdocuments.model.templates
+package net.atos.client.smartdocuments.model.template
 
 import nl.lifely.zac.util.NoArgConstructor
